Extract RowBuilder class from Row and add tests for it
