+++ lib/honey_format/row.rb
@@ -1,24 +1,8 @@
+require 'honey_format/row_builder'
+
 module HoneyFormat
   # Holds data for a single row.
   class Row
-    class RowBuilder < Struct
-      def self.call(row)
-        new(*row)
-      end
-
-      # @return [String] CSV-string representation.
-      def to_csv
-        members.map do |column_name|
-          column = public_send(column_name)
-          if column.respond_to?(:to_csv)
-            column.to_csv
-          else
-            column.to_s
-          end
-        end.join(',') + "\n"
-      end
-    end
-
     # Returns a new instance of Row.
     # @return [Row] a new instance of Row.
     # @param [Array] columns an array of symbols.
